+++ stories/index.story.jsx
@@ -56,6 +56,7 @@ stories.add('absolute (default)', () => {
     />
   );
 });
+
 stories.add('relative range', () => {
   const width = text('Width', '200px');
   const translations = {
@@ -115,6 +116,7 @@ stories.add('relative range', () => {
     />
   );
 });
+
 stories.add('period', () => {
   const width = text('Width', '200px');
   const translations = {
@@ -228,6 +230,83 @@ stories.add('all', () => {
       enabled={enabled}
       onChange={action('handleChange')}
       relativeRange={relativeRange}
+      translations={translations}
+      width={width}
+    />
+  );
+});
+
+stories.add('Range field', () => {
+  const width = text('Width', '200px');
+  const translations = {
+    startDate: text('Start date label', 'First day'),
+    endDate: text('End date label', 'Last day'),
+    absolute: text('Absolute label', 'Absolute'),
+    relative: text('Relative label', 'Relative'),
+  };
+  const absoluteEnabled = boolean('Absolute range enabled', true);
+  const periodEnabled = boolean('Period enabled', true);
+  const relativeEnabled = boolean('Relative range enabled', true);
+  const dateFormat = text('Date format', 'DD.MM.YYYY');
+  const numberOfMonths = number('Number of months', 1);
+  const region = text('Region', 'en_GB');
+  const showWeekNumbers = boolean('Show week numbers', true);
+  const periodStartDate = object(
+    'Period start date',
+    {
+      unit: 'DAY',
+      timing: 'CURRENT',
+    },
+  );
+  const periodEndDate = object(
+    'Period end date',
+    {
+      unit: 'DAY',
+      timing: 5,
+      periodic: true,
+    },
+  );
+  const enabled = {
+    absolute: absoluteEnabled,
+    period: periodEnabled,
+    relative: relativeEnabled,
+  };
+  const absoluteRange = {
+    dateFormat,
+    numberOfMonths,
+    region,
+    showWeekNumbers,
+  };
+  const period = {
+    endDate: periodEndDate,
+    startDate: periodStartDate,
+  };
+  const rangeField = {
+    onChange: action('handleChange'),
+    options: [
+      {
+        label: 'Created',
+        value: 'created',
+      },
+      {
+        label: 'Modified',
+        value: 'modified',
+      },
+    ],
+    value: {
+      label: 'Modified',
+      value: 'modified',
+    }
+  };
+
+  return (
+    <DateRange
+      id="default"
+      absoluteRange={absoluteRange}
+      enabled={enabled}
+      onChange={action('handleChange')}
+      period={period}
+      rangeField={rangeField}
       translations={translations}
       width={width}
     />